agent_framework/requirements.txt
# Core dependencies
pyyaml>=6.0
jsonschema>=4.17.3

# For RAG integration
numpy>=1.24.3
sentence-transformers>=2.2.2
transformers>=4.30.2

# Optional dependencies - uncomment as needed
# torch>=2.0.1  # Required for sentence-transformers and transformers
# bitsandbytes>=0.39.1  # For 8-bit quantization
# chromadb>=0.4.6  # If you're using ChromaDB for vector storage
# rank_bm25>=0.2.2  # For keyword search in EnhancedRAG

# For optimization
tqdm>=4.65.0

# For testing
pytest>=7.3.1